// src/Component/types/ProductType.ts
export type TFlowerProduct= {
        _id:string,
        productName: string
        price: number
        quantity: number
        bloomDate: string
        color: string[]
        type: string
        size: string
        fragrance: string
        arrangementStyle: string
        occasion: string
      }